- Install Jest, SuperTest - Create .test.js file for Jest - Move express routes to src/routes/index.js
