+++ package.json
@@ -4,7 +4,7 @@
   "description": "Backend API for Monkey Wrench Forums",
   "main": "src/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "start": "node src/index.js",
     "lint": "eslint src/*.js",
     "format": "prettier src/* --ignore-path ./.gitignore"
@@ -21,7 +21,9 @@
   "homepage": "https://github.com/khayes133/monkeywrench#readme",
   "devDependencies": {
     "eslint": "^8.43.0",
-    "prettier": "^2.8.8"
+    "jest": "^29.6.1",
+    "prettier": "^2.8.8",
+    "supertest": "^6.3.3"
   },
   "dependencies": {
     "cors": "^2.8.5",
